Fix a bug which will cause SEGV on 64-bit platform; version -> 1.2.1

+++ src/include/animation.h
@@ -59,7 +59,7 @@ struct _EffAnimation {
     int   lasttime;
     int   duration;
     int   priority;
-    int   property_id;
+    intptr_t property_id;
     void* targetobject;
     void* context;
     EffMotionCurve* curve;
